+++ jem/__init__.py
@@ -27,6 +27,7 @@ from jem.exchangers import (
     default_exchangers,
     default_exchanges,
     default_workflow,
+    exchanged_fields,
 )
 from jem.output import (
     chunk_datasets,
@@ -62,6 +63,7 @@ __all__ = [
     "default_exchanges",
     "default_health_check",
     "default_workflow",
+    "exchanged_fields",
     "nested_carry",
     "postprocess",
     "postprocess_datasets",